refactor: :truck: rename artifactid, project name, package name, and classes to sorriso
